(WIP) Single-page (single-iframe) Web app for the composer
Rather than having this clunky split UI with controls on the left and
a preview to the right, make the preview full-screen and embed the Vue
app directly within it. This requires some trickery when serving the
newsletter in edit mode, but when all is said and done, not moreso
than trying to bend the WordPress maze of callbacks to do our bidding.

* theme-options.php now has an iframe and some
  "plain" (unbrowserified) JS to get rid of the DOM set up by The
  Newsletter Plugin for us
* New handler in new class method NewsletterHook::serve_composer_app;
  calls theme.php after defining a render_editor_scripts() function
  (which theme.pp is supposed to check for)
* The plan is for the Vue controls to install themselves inside the
  newsletter's DOM wherever they want, and reload the page upon every
  state change