Bump memory limit
The import of the "big" news.json always fails around import_id=1900
or so (out of 2183 as of this writing); sometimes (though not always)
mentioning a memory exhaustion problem (which is hardly surprising
given how the typical C program, among which PHP itself, checks
malloc() return values - Or more precisely, doesn't). The current
limit appears to be 128Mb; this should buy us some headroom.